Using dp instead of sp for text sizes
When setting text sizes, you should normally use sp, or
"scale-independent pixels". This is like the dp unit, but it is also
scaled by the user's font size preference. It is recommend you use
this unit when specifying font sizes, so they will be adjusted for both
the screen density and the user's preference.